requirements.txt
streamlit
pandas
numpy
matplotlib
openai
tiktoken
st-supabase-connection
chardet
PyYAML
memory-profiler
pillow